requirements.txt
# --- BACKEND BASE (FastAPI + MinIO) ---
fastapi
uvicorn

# --- MODELLI TTS ---

# 1. PIPER
piper-tts

# 2. KOKORO 
kokoro-onnx==0.4.9
soundfile
onnxruntime

# 3. BARK 
git+https://github.com/suno-ai/bark.git

# 4. COQUI in ambiente indipendente
#TTS==0.14.3

# --- DIPENDENZE EXTRA AUDIO ---
scipy
numpy==2.2.6